requirements.txt
crewai
langchain_google_genai #Modelos do Gemini do Google
load_dotenv
crewai_tools